Move SDMX-JSON messages below the io module
