+++ src/pysdmx/io/json/sdmxjson2/messages/core.py
