+++ src/pysdmx/io/json/sdmxjson2/messages/dataflow.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Sequence
 
 from msgspec import Struct
 
-from pysdmx.fmr.sdmx.org import JsonDataProviderScheme
+from pysdmx.io.json.sdmxjson2.messages.org import JsonDataProviderScheme
 from pysdmx.model import (
     Agency,
     Components,
