+++ src/pysdmx/io/json/sdmxjson2/messages/dsd.py
@@ -4,9 +4,12 @@ from typing import Dict, List, Optional, Sequence, Tuple
 
 from msgspec import Struct
 
-from pysdmx.fmr.sdmx.concept import JsonConcept, JsonConceptScheme
-from pysdmx.fmr.sdmx.constraint import JsonContentConstraint
-from pysdmx.fmr.sdmx.core import JsonRepresentation
+from pysdmx.io.json.sdmxjson2.messages.concept import (
+    JsonConcept,
+    JsonConceptScheme,
+)
+from pysdmx.io.json.sdmxjson2.messages.constraint import JsonContentConstraint
+from pysdmx.io.json.sdmxjson2.messages.core import JsonRepresentation
 from pysdmx.model import (
     ArrayBoundaries,
     Codelist,
